Factor out vote processor
